pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "discord-audio-router"
version = "2.0.0"
description = "Professional Discord bot for real-time audio routing between voice channels"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Discord Audio Router Team", email = "team@discord-audio-router.com"}
]
maintainers = [
    {name = "Discord Audio Router Team", email = "team@discord-audio-router.com"}
]
keywords = ["discord", "bot", "audio", "routing", "voice", "websocket", "broadcast", "streaming"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
requires-python = ">=3.9"
dependencies = [
    "discord.py==2.4.0",
    "discord-ext-voice-recv>=0.4.2a145",
    "websockets>=10.4",
    "aiohttp>=3.8.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0",
    "numpy>=1.21.0",
    "ffmpeg-python>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
    "pre-commit>=2.20.0",
]
docs = [
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "myst-parser>=0.18.0",
]

[project.urls]
Homepage = "https://github.com/discord-audio-router/discord-audio-router"
Documentation = "https://discord-audio-router.readthedocs.io/"
Repository = "https://github.com/discord-audio-router/discord-audio-router.git"
Issues = "https://github.com/discord-audio-router/discord-audio-router/issues"

[project.scripts]
discord-audio-router = "discord_audio_router.bots.main_bot:main"
discord-audio-forwarder = "discord_audio_router.bots.forwarder_bot:main"
discord-audio-receiver = "discord_audio_router.bots.receiver_bot:main"
discord-audio-relay = "discord_audio_router.networking.websocket_server:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["discord_audio_router*"]

[tool.setuptools.package-data]
"discord_audio_router" = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["discord_audio_router"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "discord.*",
    "discord.ext.*",
    "websockets.*",
    "numpy.*",
    "ffmpeg.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

